% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipRadius.R
\name{getStates}
\alias{getStates}
\title{getStates returns the list of states which have zip codes which fall in a specified radius in lower case format for use in choroplethrZip}
\usage{
getStates(zipcode, radius)
}
\arguments{
\item{zipcode}{the reference zip code of which you'd like the list of zip codes within a give radius as character}

\item{radius}{the distance in miles from the center of the given zip to the center of the other zips as numeric}
}
\description{
getStates returns the list of states which have zip codes which fall in a specified radius in lower case format for use in choroplethrZip
}
\examples{
getStates("30316", 10)

}
