% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipRadius.R
\name{lookupZips}
\alias{lookupZips}
\title{Lookup zip code location}
\usage{
lookupZips(zipcodes)
}
\arguments{
\item{zipcodes}{A character vector of five digit zip codes}
}
\value{
A data frame with five columns and as many rows as
the length of the \code{zipcodes} argument.
\describe{
\item{zip}{Character. The zip codes you passed to the function.}
\item{city}{Character. The cities containing the zip codes.}
\item{state}{Character. The states containing the zip codes.}
\item{latitude}{Numeric. The latitudes of the centers of the zip codes.}
\item{longitude}{Numeric. The longitudes of the centers of the zip codes.}
}
}
\description{
Get the city, state, latitude and longitude of zip codes
}
\examples{
zipcodes <- c("99501", "90210", "33162", "60606", "42748", "99999")
lookupZips(zipcodes)
}
