% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{zipList}
\alias{zipList}
\title{zipList}
\format{A data frame with five columns and 44336 rows.
\describe{
\item{zip}{Character. Five-digit US zip codes.}
\item{city}{Character. The cities containing the zip codes.}
\item{state}{Character. The states containing the zip codes.}
\item{latitude}{Numeric. The latitudes of the centers of the zip codes.}
\item{longitude}{Numeric. The longitudes of the centers of the zip codes.}
}}
\description{
US zip codes, and their city, state, latitude, and longitude.
}
