% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipRadius.R
\name{getZipPop}
\alias{getZipPop}
\title{get the population living near a zip code}
\usage{
getZipPop(zipcode, radius)
}
\arguments{
\item{zipcode}{The reference zip code of which you'd like the list of zip
codes within a give radius as character.}

\item{radius}{The distance in miles from the center of the given zip to the
center of the other zips as numeric.}
}
\value{
A data frame with two columns.
the length of the \code{zipcodes} argument.
\describe{
\item{region}{Character. A zip code nearby to the reference zip code.}
\item{value}{Numeric. The population of that zip code.}
}
}
\description{
\code{getZipPop} returns a data frame of zip codes and their population where
 the zip codes fall within a given radius for use in \code{choroplethrZip}.
}
\examples{
getZipPop("30316", 10)
}
