% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipRadius.R
\name{zipRadius}
\alias{zipRadius}
\title{The zipRadius Function}
\usage{
zipRadius(zipcode, radius)
}
\arguments{
\item{zipcode}{the reference zip code of which you'd like the list of zip codes within a give radius as character}

\item{radius}{the distance in miles from the center of the given zip to the center of the other zips as numeric}
}
\description{
The zipRadius Function
}
\examples{
zipRadius("30316", 10)
}
